% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.configs.SIPNET.R
\name{binary_args_to_string}
\alias{binary_args_to_string}
\title{Convert binary_args list to command-line argument string.}
\usage{
binary_args_to_string(args)
}
\description{
Accepts a list like list(arg1 = "value", arg2 = 3, arg3 = NULL),
which becomes "--arg1=value --arg2=3 --arg3".
If provided as a plain string, passes through with a warning.
}
