% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sipnet_out.R
\name{combine_sipnet_out}
\alias{combine_sipnet_out}
\title{Combine a bunch of \code{sipnet.out} files into a single file}
\usage{
combine_sipnet_out(directory, outfile, files = NULL)
}
\arguments{
\item{directory}{Parent directory to search for \code{sipnet.out} files. You must
provide either this or an explicit vector of \code{files}.}

\item{outfile}{File to which combined \code{sipnet.out} will be written}

\item{files}{(optional) Explicit vector of paths to \code{sipnet.out} files to
combine. Note that files need not be named \code{sipnet.out}, but they must be
readable with \code{\link[=read_sipnet_out]{read_sipnet_out()}}.}
}
\value{
\code{outfile} (path to output file), invisibly
}
\description{
Combine a bunch of \code{sipnet.out} files into a single file
}
