% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_string_templates.R
\name{expand_string_templates}
\alias{expand_string_templates}
\title{Replace \verb{@TEMPLATE@} placeholders in a set of lines of text}
\usage{
expand_string_templates(
  text,
  settings,
  ...,
  HOST_SETUP = paste(paste(settings$model$prerun, collapse = "\\n"),
    paste(settings$host$prerun, collapse = "\\n"), "", sep = "\\n"),
  CDO_SETUP = paste(settings$host$cdosetup, "", collapse = "\\n"),
  HOST_TEARDOWN = paste(paste(settings$model$postrun, collapse = "\\n"),
    paste(settings$host$postrun, collapse = "\\n"), "", sep = "\\n"),
  CPRUNCMD = "",
  CPOUTCMD = "",
  RMOUTDIRCMD = "",
  RMRUNDIRCMD = "",
  SITE_LAT = settings$run$site$lat,
  SITE_LON = settings$run$site$lon,
  SITE_MET = settings$run$inputs$met$path,
  OUTDIR = shQuote(settings$host$outdir),
  RUNDIR = shQuote(settings$host$rundir),
  START_DATE = settings$run$start.date,
  END_DATE = settings$run$end.date,
  BINARY = settings$model$binary,
  REVISION = settings$model$revision,
  PREFIX = settings$state.data.assimilation$NC.Prefix,
  OVERWRITE = settings$state.data.assimilation$NC.Overwrite,
  CONFLICT = settings$state.data.assimilation$FullYearNC,
  DELETE.RAW = settings$model$delete.raw
)
}
\arguments{
\item{text}{lines of text (destined for job.sh) in which to replace @strings@}

\item{settings}{PEcAn settings object to look in for default values}

\item{...}{other templates to replace, listed first so they can include
later-expanded templates within themselves}

\item{HOST_SETUP, CDO_SETUP, HOST_TEARDOWN}{host-specific setup and teardown}

\item{CPRUNCMD, CPOUTCMD, RMOUTDIRCMD, RMRUNDIRCMD}{commands to manipulate and
clean up the workflow directory after model execution. Not all templates
use these}

\item{SITE_LAT, SITE_LON, START_DATE, END_DATE}{location and date range being
modeled}

\item{SITE_MET}{path to clim file}

\item{OUTDIR, RUNDIR}{PEcAn run and out directories for this invocation}

\item{BINARY, REVISION}{Sipnet model information}

\item{PREFIX, OVERWRITE, CONFLICT}{settings specific to SDA workflow}

\item{DELETE.RAW}{logical: Delete each \code{sipnet.out} after writing to netcdf?}
}
\value{
Modified text with the listed @TEMPLATE@s replaced
}
\description{
Helper that mostly exists to save (a little) typing when copying string
template replacement from write.configs into write_segmented_configs.
}
\details{
Replacements are done iteratively in argument order and all apply to the
entire text, so earlier-listed values can contain templates for later
expansion, e.g.
\code{expand_string_templates(txt, settings, SETUP="mkdir @MYDIR@", MYDIR="out")}
will expand "bash -c '@SETUP@'" to "bash -c 'mkdir out'".
}
\author{
Chris Black
}
\keyword{internal}
