% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.model.output.SIPNET.R
\name{get.model.output.SIPNET}
\alias{get.model.output.SIPNET}
\title{Retrieve model output from local or remote server}
\usage{
get.model.output.SIPNET(settings)
}
\description{
Function to retrieve model output from local or remote server
}

