% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.SIPNET.R
\name{mergeNC}
\alias{mergeNC}
\title{Merge multiple NetCDF files into one}
\source{
https://github.com/RS-eco/processNC/blob/main/R/mergeNC.R
}
\usage{
mergeNC(files, outfile)
}
\arguments{
\item{files}{\code{character}. List of filepaths, which should lead to
NetCDF files.}

\item{outfile}{\code{character}. Output filename of the merged data.}
}
\value{
A NetCDF file containing all of the merged data.
}
\description{
Merge multiple NetCDF files into one
}
\examples{
\dontrun{
files <- list.files(paste0(system.file(package = "processNC"), "/extdata"),
  pattern = "tas.*\\\\.nc", full.names = TRUE
)
temp <- tempfile(fileext = ".nc")
mergeNC(files = files, outfile = temp)
terra::rast(temp)
}
}
