% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.SIPNET.R
\name{model2netcdf.SIPNET}
\alias{model2netcdf.SIPNET}
\title{Convert SIPNET output to netCDF}
\usage{
model2netcdf.SIPNET(
  outdir,
  sitelat,
  sitelon,
  start_date,
  end_date,
  delete.raw = FALSE,
  revision,
  prefix = "sipnet.out",
  overwrite = FALSE,
  conflict = TRUE
)
}
\arguments{
\item{outdir}{Location of SIPNET model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}

\item{delete.raw}{logical: remove sipnet.out files after converting?}

\item{revision}{model revision}

\item{prefix}{prefix to read the output files}

\item{overwrite}{Flag for overwriting nc files or not}

\item{conflict}{Flag for dealing with conflicted nc files, if T we then will merge those, if F we will jump to the next.}
}
\description{
Converts all output contained in a folder to netCDF.
}
\author{
Shawn Serbin, Michael Dietze
}
