% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.SIPNET.R
\name{parse_sipnet_datetime}
\alias{parse_sipnet_datetime}
\title{Convert SIPNET output to netCDF}
\usage{
parse_sipnet_datetime(x)
}
\arguments{
\item{outdir}{Location of SIPNET model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}

\item{revision}{model revision.
Ignored: PEcAn detects all relevant version differences from the format of the output file.}

\item{overwrite}{Flag for overwriting nc files or not}

\item{conflict}{Flag for dealing with conflicted nc files, if T we then will merge those, if F we will jump to the next.}

\item{prefix}{prefix to read the output files}

\item{delete.raw}{logical: remove sipnet.out files after converting?}
}
\description{
Converts all output contained in a folder to netCDF.
}
\author{
Shawn Serbin, Michael Dietze
}
