% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sipnet_out.R
\name{read_sipnet_out}
\alias{read_sipnet_out}
\title{Read \code{sipnet.out} file to \code{data.frame}}
\usage{
read_sipnet_out(sipnet_out_file)
}
\arguments{
\item{sipnet_out_file}{Path to \code{sipnet.out} file}
}
\value{
\code{data.frame} of SIPNET output
}
\description{
Read \code{sipnet.out} file to \code{data.frame}
}
