% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remove.config.SIPNET}
\alias{remove.config.SIPNET}
\title{Clear out previous SIPNET config and parameter files.}
\usage{
remove.config.SIPNET(main.outdir, settings)
}
\arguments{
\item{main.outdir}{Primary PEcAn output directory (will be depreciated)}

\item{settings}{PEcAn settings file}
}
\value{
nothing, removes config files as side effect
}
\description{
Clear out previous SIPNET config and parameter files.
}
\author{
Shawn Serbin, David LeBauer
}

