% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.SIPNET.R
\name{sipnet2datetime}
\alias{sipnet2datetime}
\title{Convert SIPNET DOY to datetime}
\usage{
sipnet2datetime(sipnet_tval, base_year, base_month = 1,
  force_cf = FALSE)
}
\arguments{
\item{sipnet_tval}{vector of SIPNET DOY values}

\item{base_year}{base year to calculate datetime from DOY}

\item{base_month}{reference month for converting from DOY to datetime}

\item{force_cf}{force output to follow CF convention. Default FALSE}
}
\description{
Convert SIPNET DOY to datetime
}
\author{
Alexey Shiklomanov, Shawn Serbin
}
