% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sipnet_time_range_idx.R
\name{sipnet_time_range_idx}
\alias{sipnet_time_range_idx}
\title{Compute first and last indices to extract from a Sipnet clim file}
\usage{
sipnet_time_range_idx(
  yrs,
  ydays,
  hrs,
  start_yr,
  start_yday,
  start_hr,
  end_yr,
  end_yday,
  end_hr
)
}
\arguments{
\item{yrs, ydays, hrs}{vectors to be subset,
containing respectively the year, day-of-year, and hour of day.}

\item{start_yr, start_yday, start_hr}{year, day-of-year, and hour
to start from, all scalar}

\item{end_yr, end_yday, end_hr}{year, day-of-year, and hour
to stop \emph{before}, all scalar}
}
\description{
Given vectors of integer years, integer days-of-year, and decimal hours,
computes the first and last indices that fall between the given start
and end times.
}
\details{
Assumes the inputs are sorted (as Sipnet requires), and specifically that
the first and last rows are the earliest and latest timepoints.
If they are not, the result will be unpredictable.
}
\author{
Chris Black
}
