% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_inputs.SIPNET.R
\name{split_inputs.SIPNET}
\alias{split_inputs.SIPNET}
\title{Split SIPNET inputs into multiple files based on start and end time}
\usage{
split_inputs.SIPNET(
  start.time,
  stop.time,
  inputs,
  overwrite = FALSE,
  outpath = NULL,
  time_resolution = c("row", "day"),
  allow_varying_timesteps = TRUE
)
}
\arguments{
\item{start.time}{Start date or datetime for splitting}

\item{stop.time}{End date or datetime for splitting}

\item{inputs}{Named \code{inputs} list as provided by PEcAn \code{settings}. Must have
structure like:
\code{list(met = list(path = "path/to/sipnet.clim"), events = list(path = "path/to/events.in"), ...)}}

\item{overwrite}{if \code{TRUE}, overwrite existing target file (Default \code{FALSE})}

\item{outpath}{if specified, write output to a new directory. Default \code{NULL} writes back to the directory being read}

\item{time_resolution}{One of "row", "day". Should the split happen
at the closest row to the end time or align with a time unit? See details}

\item{allow_varying_timesteps}{logical: Accept files where not all rows have
the same duration? Disallowing can be substantially faster.}
}
\value{
Modified \code{inputs} list with all \code{path} entries replaced with new
paths (suitable for inserting back into \code{settings$run$inputs}).
}
\description{
Subset each SIPNET input file and write a new file containing values \verb{>= start.time} and \verb{< end.time} (note: end time is \emph{not} inclusive)
}
\details{
NOTE that sipnet met files contain dates \emph{and} times, while sipnet event
files contain only dates. Comparing a datetime to a date will coerce the
date to midnight UTC.
}
\section{Efficiency}{


Splitting met files goes faster by far when all the inputs are known to have
a uniform timestep, fixed number of characters per line, and all splits are
at day boundaries. Setting \code{time_resolution = "days"} and
\code{allow_varying_timesteps = FALSE} opts in to a fast code path that finds the
wanted lines of the file by computing byte offsets and therefore avoids
needing to read every row of the input file.
For small subsets of large files this can be several times faster.

If the input has variable line length but constant timesteps (as determined
by checking the first 10 lines; see the \code{n_head} argument of
\code{subset_met_lines()}), the fast path is taken but is not \emph{as} fast: it
must scan from the beginning to count line separators, but still skips the
tail of the file after \code{stop.time} is reached.
}

\author{
Alexey Shiklomanov
}
