% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_inputs.SIPNET.R
\name{split_sipnet_events}
\alias{split_sipnet_events}
\title{Split sipnet \code{events.in} files according to start and stop date}
\usage{
split_sipnet_events(
  start.time,
  stop.time,
  eventfile,
  overwrite = FALSE,
  outpath = FALSE
)
}
\arguments{
\item{start.time}{Start date or datetime for splitting}

\item{stop.time}{End date or datetime for splitting}

\item{eventfile}{Path to \code{events.in} file.}

\item{overwrite}{if \code{TRUE}, overwrite existing target file (Default \code{FALSE})}

\item{outpath}{if specified, write output to a new directory. Default \code{NULL} writes back to the directory being read}
}
\description{
Split sipnet \code{events.in} files according to start and stop date
}
