% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/start.runs.SIPNET.R
\name{start.runs.SIPNET}
\alias{start.runs.SIPNET}
\title{Start SIPNET model runs}
\usage{
\method{start}{runs.SIPNET}(runid)
}
\arguments{
\item{runid}{the id of the run that needs to be executed}
}
\description{
Start SIPNET model runs on local server
}
\author{
Michael Dietze, David LeBauer, Shawn Serbin, Carl Davidson
}

