% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_segmented_configs.R
\name{subset_paths}
\alias{subset_paths}
\title{settings for whole run with many paths per input ->
settings for one ensemble member with one path per input}
\usage{
subset_paths(settings, path_nums)
}
\arguments{
\item{settings}{single-site settings object (not Multisettings)}

\item{path_nums}{named list of input indices (one row of the sample design)}
}
\description{
settings for whole run with many paths per input ->
settings for one ensemble member with one path per input
}
