% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg2model.SIPNET.R
\name{veg2model.SIPNET}
\alias{veg2model.SIPNET}
\title{veg2model.SIPNET}
\usage{
veg2model.SIPNET(input, outdir, siteid, ens)
}
\arguments{
\item{input}{input object taken from cohort2pool}

\item{outdir}{location where you want ncdf files stored}

\item{siteid}{BETY site id}

\item{ens}{number of ensemble members}
}
\value{
result object with filepaths to ncdf files
}
\description{
veg2model.SIPNET
}
\author{
Alexis Helgeson
}
