% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.restart.SIPNET.R
\name{write.restart.SIPNET}
\alias{write.restart.SIPNET}
\title{write.restart.SIPNET}
\usage{
write.restart.SIPNET(outdir, runid, start.time, stop.time, settings, new.state,
  RENAME = TRUE, new.params = FALSE, inputs)
}
\arguments{
\item{outdir}{output directory}

\item{runid}{run ID}

\item{settings}{PEcAn settings object}

\item{new.state}{analysis state vector}

\item{RENAME}{flag to either rename output file or not}

\item{time}{year that is being read}

\item{met}{}
}
\value{
NONE
}
\description{
Write restart files for SIPNET
}
\author{
Ann Raiho \email{araiho@nd.edu}
}

