% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_segmented_configs.R
\name{write_segmented_configs.SIPNET}
\alias{write_segmented_configs.SIPNET}
\title{Break a Sipnet simulation into time segments with possibly changed parameters}
\usage{
write_segmented_configs.SIPNET(settings, input_design = NULL, ...)
}
\arguments{
\item{settings}{a single PEcAn settings object (not MultiSettings)}

\item{input_design}{data frame in the format returned by
\code{PEcAn.uncertainty::generate_joint_ensemble_design()},
reporting which ensemble members to select for each input}

\item{...}{further arguments passed on to \code{write_segment_configs()}}
}
\description{
Each segment will run in its own subdirectory,
starting from the restart file written at the end of the previous segment.
Crop parameters can be altered by specifying a different PFT in the segment
file.
}
\author{
Alexey Shiklomanov, Chris Black
}
