% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_restarts.R
\name{collect_restarts}
\alias{collect_restarts}
\title{Copy final model state from all sites in a completed PEcAn workflow}
\usage{
collect_restarts(run_dir, dest_dir, overwrite = FALSE)
}
\arguments{
\item{run_dir}{path to the run directory of a PEcAn workflow.
This should be the directory listed in the workflow's \code{settings$rundir};
by convention this often ends in \verb{output/run/}}

\item{dest_dir}{directory to which to copy one restart file per model run}

\item{overwrite}{logical: Replace existing destination files?}
}
\value{
Invisibly, a character vector of the filenames created in dest_dir.
}
\description{
Finds the \code{restart.out} files in a PEcAn run directory, copies the last
restart from each run to the specified output directory,
and renames by site and ensemble ID.
This is useful when taking final state from one simulation as the starting
point for another one (e.g. forecasts that start from a current-observations
run).
}
\details{
Paths look like either \code{ENS-00001-131976/restart.out} (in runs that are all
one segment) or \code{ENS-00004-131975/segments/segment_011/run/restart.out}
(in segmented runs).
Segmented runs will have a restart.out for each segment, in which case
we take the restart from the last segment in the run.
(Well, technically the one whose path sorts last, with our confidence that
this is also chonologically last coming from the segment numbering system).

On copying, files are renamed by site ID and ensemble member. Segment number
is not retained. Thus the examples shown above would respectively become
\verb{<dest_dir>/restart-131976-00001.out} and
\verb{<dest_dir>/restart-131975-00004.out}.
}
\author{
Chris Black
}
