% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.SIPNET.R
\name{met2model.SIPNET}
\alias{met2model.SIPNET}
\title{met2model wrapper for SIPNET}
\usage{
met2model.SIPNET(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  var.names = NULL,
  overwrite = FALSE,
  verbose = FALSE,
  year.fragment = FALSE,
  add_grid_loc = FALSE,
  ...
)
}
\arguments{
\item{in.path}{location on disk where inputs are stored}

\item{in.prefix}{prefix of input and output files,}

\item{outfolder}{location on disk where outputs will be stored}

\item{start_date}{the start date of the data to be downloaded
(will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded
(will only use the year part of the date)}

\item{var.names}{character: list of variable names to be extracted. Default is NULL.}

\item{overwrite}{should existing files be overwritten}

\item{verbose}{should the function be very verbose}

\item{year.fragment}{the function should ignore whether or not the data is
stored as a set of complete years (such as for forecasts).}

\item{add_grid_loc}{Add column of zeroes as an \link{unused!} spatial grid index?
Provided for backwards compatibility; Set to TRUE if you need the output to
be readable by Sipnet versions older than v2.0.}

\item{...}{Additional arguments, currently ignored}
}
\value{
a dataframe containing information about the files created
}
\description{
Reads weather data from CF-formatted NetCDFs and writes it in the \code{.clim}
format expected by SIPNET: a 13-column tab-separated table with no headers.
}
\details{
The columns of the output file are:
\itemize{
\item 4-digit year
\item Day of year
\item Hour of day
\item Timestep size (days)
\item Air temperature (degrees C)
\item Soil temperature (degrees C)
\item PAR (mol/m2/hr)
\item Precip (mm)
\item VPD (Pa)
\item VPD of soil (Pa)
\item Canopy vapor pressure (Pa)
\item Wind speed (m/s)
\item Soil moisture (fraction of saturation).
Always  0.6 from this function; PEcAn configures SIPNET to calculate it
internally.
}

If add_grid_loc is TRUE, the table has 14 columns: The 13 above,
plus a new leftmost dummy column containing all zeroes.
This is a grid location index that was needed by older versions of Sipnet
and is ignored with a warning by new versions.
Note that PEcAn has never used gridded simulations, and model support for
them was removed from Sipnet >= v2.0.

SIPNET does not allow missing values in its inputs. If the result contains
NAs after conversion, no file is written and the process returns an error.
To fix this, consider using a formal gap-filling method such as
\code{PEcAn.data.atmosphere::metgapfill()} before calling met2model.
}
\author{
Luke Dramko, Michael Dietze, Alexey Shiklomanov, Rob Kooper
}
