% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.SIPNET.R
\name{met2model.SIPNET}
\alias{met2model.SIPNET}
\title{met2model wrapper for SIPNET}
\usage{
met2model.SIPNET(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE,
  verbose = FALSE,
  year.fragment = FALSE,
  ...
)
}
\arguments{
\item{in.path}{location on disk where inputs are stored}

\item{in.prefix}{prefix of input and output files,}

\item{outfolder}{location on disk where outputs will be stored}

\item{start_date}{the start date of the data to be downloaded
(will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded
(will only use the year part of the date)}

\item{overwrite}{should existing files be overwritten}

\item{verbose}{should the function be very verbose}

\item{year.fragment}{the function should ignore whether or not the data is
stored as a set of complete years (such as for forecasts).}

\item{...}{Additional arguments, currently ignored}
}
\value{
a dataframe containing information about the files created
}
\description{
Reads weather data from CF-formatted NetCDFs and writes it in the \code{.clim}
format expected by SIPNET: a 14-column tab-separated table with no headers.
}
\details{
The columns of the output file are:
\itemize{
\item Grid index. Always 0 from this function; PEcAn configures SIPNET not to
use this.
\item 4-digit year
\item Day of year
\item Hour of day
\item Timestep size (days)
\item Air temperature (degrees C)
\item Soil temperature (degrees C)
\item PAR (mol/m2/hr)
\item Precip (mm)
\item VPD (Pa)
\item VPD of soil (Pa)
\item Canopy vapor pressure (Pa)
\item Wind speed (m/s)
\item Soil moisture (fraction of saturation).
Always  0.6 from this function; PEcAn configures SIPNET to calculate it
internally.
}

Input rows containing NAs in any column are removed with a warning.
This ensures the output is readable by SIPNET (which does not allow missing
values in its inputs), but does not do any correction for the resulting
gaps in the time series. If your file contains more than incidental missing
values, consider using a formal gap-filling method before calling met2model.
}
\author{
Luke Dramko, Michael Dietze, Alexey Shiklomanov, Rob Kooper
}
