% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.SIPNET.R
\name{met2model.SIPNET}
\alias{met2model.SIPNET}
\title{met2model.SIPNET}
\usage{
met2model.SIPNET(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE,
  verbose = FALSE,
  year.fragment = FALSE,
  ...
)
}
\arguments{
\item{in.path}{location on disk where inputs are stored}

\item{in.prefix}{prefix of input and output files OR the full file name if year.fragment = TRUE}

\item{outfolder}{location on disk where outputs will be stored}

\item{start_date}{the start date of the data to be downloaded (will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded (will only use the year part of the date)}

\item{overwrite}{should existing files be overwritten}

\item{verbose}{should the function be very verbose}

\item{year.fragment}{the function should ignore whether or not the data is stored as a set of complete years (such as for forecasts).}

\item{...}{Additional arguments, currently ignored}
}
\description{
met2model wrapper for SIPNET
}
\author{
Luke Dramko, Michael Dietze, Alexey Shiklomanov, Rob Kooper
}
