% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.SIPNET.R
\name{model2netcdf.SIPNET}
\alias{model2netcdf.SIPNET}
\title{Function to convert SIPNET model output to standard netCDF format}
\usage{
model2netcdf.SIPNET(outdir, sitelat, sitelon, start_date, end_date,
  delete.raw, revision, overwrite = FALSE, obs.t = NULL)
}
\arguments{
\item{outdir}{Location of SIPNET model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}

\item{revision}{model revision}

\item{overwrite}{Flag for overwriting nc files or not}
}
\description{
Convert SIPNET output to netCDF
}
\details{
Converts all output contained in a folder to netCDF.
}
\author{
Shawn Serbin, Michael Dietze
    model2netcdf.SIPNET('/data/bmorrison/sda/lai/4_sites_8_days/out/ENS-00001-676', 45.805925, -90.07961, '1980-01-01', '2009-12-31', FALSE, 'r136')
}
