\name{model2netcdf.SIPNET}
\alias{model2netcdf.SIPNET}
\title{Function to convert SIPNET model output to standard netCDF format}
\usage{
  model2netcdf.SIPNET(outdir, run.id)
}
\arguments{
  \item{outdir}{Location of SIPNET model output}

  \item{run.id}{Name of SIPNET model output file.}
}
\description{
  Function to convert SIPNET model output to standard
  netCDF format
}
\author{
  Shawn Serbin, Michael Dietze
}

