\name{model2netcdf.SIPNET}
\alias{model2netcdf.SIPNET}
\title{Function to convert SIPNET model output to standard netCDF format}
\usage{
  model2netcdf.SIPNET(outdir)
}
\arguments{
  \item{outdir}{Location of SIPNET model output}
}
\description{
  Convert SIPNET output to netCDF
}
\details{
  Converts all output contained in a folder to netCDF.
}
\author{
  Shawn Serbin, Michael Dietze
}

