\name{read.restart.SIPNET}
\alias{read.restart.SIPNET}
\title{read.restart.SIPNET}
\usage{
read.restart.SIPNET(outdir, runid, time, settings, variables,
  sample_parameters = NULL)
}
\arguments{
  \item{outdir}{output directory} \item{runid}{run ID}
  \item{time}{year that is being read} \item{settings}{PEcAn
  settings object} \item{variables}{variables to be extracted}
}
\value{
X.vec vector of forecasts
}
\description{
Read Restart for SIPNET
}
\author{
Ann Raiho \email{araiho@nd.edu}
}

