% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.restart.SIPNET}
\alias{read.restart.SIPNET}
\title{read.restart.SIPNET}
\usage{
read.restart.SIPNET(outdir, runid, stop.time, settings, var.names,
  sample_parameters = NULL)
}
\arguments{
\item{outdir}{output directory}

\item{runid}{run ID}

\item{stop.time}{year that is being read}

\item{settings}{PEcAn settings object}

\item{var.names}{var.names to be extracted}
}
\value{
X.vec      vector of forecasts
}
\description{
Read Restart for SIPNET
}
\author{
Ann Raiho \email{araiho@nd.edu}
}

