% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_restart.SIPNET.R
\name{read_restart.SIPNET}
\alias{read_restart.SIPNET}
\title{Read restart function for SDA with SIPNET}
\usage{
read_restart.SIPNET(outdir, runid, stop.time, settings, var.names, params)
}
\value{
X.vec      vector of forecasts
}
\description{
Read Restart for SIPNET
}
\author{
Ann Raiho \email{araiho@nd.edu}
}

