% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_restart.SIPNET.R
\name{read_restart.SIPNET}
\alias{read_restart.SIPNET}
\title{Read restart function for SDA with SIPNET}
\usage{
read_restart.SIPNET(outdir, runid, stop.time, settings, var.names, params,
  var.map = list(c("leaf_carbon_content", "litter_carbon_content",
  "TotSoilCarb", "SoilMoistFrac", "SWE", "GWBI", "AbvGrndWood", "AGB"),
  c("LeafC", "Litter", "TotSoilCarb", "SoilMoistFrac", "SWE", "GWBI",
  "AbvGrndWood", "AGB"), c("", "", "", "", "", "kg/m^2/s", "kg/m^2", ""),
  c("", "", "", "", "", "Mg/ha/yr", "Mg/ha", ""), c("", "", "", "", "",
  mean, "", "")), timez = "UTC", When = NULL, control = list(trace =
  F), ...)
}
\arguments{
\item{outdir}{Output directory}

\item{runid}{Run ID}

\item{stop.time}{Year that is being read}

\item{settings}{PEcAn settings object}

\item{var.names}{Variable names to be extracted}

\item{params}{Any parameters required for state calculations}
}
\value{
X.vec      vector of forecasts
}
\description{
Read Restart for SIPNET
}
\author{
Ann Raiho \email{araiho@nd.edu}, Hamze Dokoohaki \email{hamzed@bu.edu}
}
