% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.IC.SIPNET.R
\name{sample.IC.SIPNET}
\alias{sample.IC.SIPNET}
\title{Sample initial conditions for SIPNET}
\usage{
sample.IC.SIPNET(ne, state, year = 1)
}
\arguments{
\item{ne}{number of ensembles}

\item{state}{state variables you want to pull}

\item{year}{year to pull from}
}
\value{
IC matrix of initial conditions
}
\description{
Sample initial conditions for SIPNET
}
\author{
Mike Dietze and Ann Raiho
}
