% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.inputs.SIPNET.R
\name{split.inputs.SIPNET}
\alias{split.inputs.SIPNET}
\title{split.inputs.SIPNET}
\usage{
\method{split}{inputs.SIPNET}(settings, start.time, stop.time)
}
\value{
file split up climate file
}
\description{
Splits climate met for SIPNET
}
\author{
Mike Dietze and Ann Raiho
}

