% Generated by roxygen2 (4.0.2): do not edit by hand
\name{split.inputs.SIPNET}
\alias{split.inputs.SIPNET}
\title{split.inputs.SIPNET}
\usage{
split.inputs.SIPNET(settings, start.time, stop.time)
}
\arguments{
\item{start.time}{}

\item{stop.time}{}

\item{multi.settings}{}
}
\value{
file split up climate file
}
\description{
Splits climate met for SIPNET
}
\author{
Mike Dietze and Ann Raiho
}

