% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_inputs.SIPNET.R
\name{split_inputs.SIPNET}
\alias{split_inputs.SIPNET}
\title{Split SIPNET inputs into multiple files based on start and end time}
\usage{
split_inputs.SIPNET(
  start.time,
  stop.time,
  inputs,
  overwrite = FALSE,
  outpath = NULL
)
}
\arguments{
\item{start.time}{Start date or datetime for splitting}

\item{stop.time}{End date or datetime for splitting}

\item{inputs}{Named \code{inputs} list as provided by PEcAn \code{settings}. Must have
structure like:
\code{list(met = list(path = "path/to/sipnet.clim"), events = list(path = "path/to/events.in"), ...)}}

\item{overwrite}{if \code{TRUE}, overwrite existing target file (Default \code{FALSE})}

\item{outpath}{if specified, write output to a new directory. Default \code{NULL} writes back to the directory being read}
}
\value{
Modified \code{inputs} list with all \code{path} entries replaced with new
paths (suitable for inserting back into \code{settings$run$inputs}).
}
\description{
Subset each SIPNET input file and write a new file containing values \verb{>= start.time} and \verb{< end.time} (note: end time is \emph{not} inclusive)
}
\details{
NOTE that sipnet met files contain dates \emph{and} times, while sipnet event
files contain only dates. Comparing a datetime to a date will coerce the
date to midnight UTC.
}
\author{
Alexey Shiklomanov
}
