% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_inputs.SIPNET.R
\name{split_inputs.SIPNET}
\alias{split_inputs.SIPNET}
\title{split_inputs.SIPNET}
\usage{
split_inputs.SIPNET(settings, start.time, stop.time, inputs)
}
\arguments{
\item{ens}{ensemble number. default = 1}
}
\value{
file split up climate file
}
\description{
Splits climate met for SIPNET
}
\author{
Mike Dietze and Ann Raiho
}

