% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_inputs.SIPNET.R
\name{split_inputs.SIPNET}
\alias{split_inputs.SIPNET}
\title{split_inputs.SIPNET}
\usage{
split_inputs.SIPNET(settings, start.time, stop.time, inputs,
  overwrite = FALSE, outpath = NULL)
}
\arguments{
\item{settings}{PEcAn settings object}

\item{start.time}{start date and time for each SDA ensemble}

\item{stop.time}{stop date and time for each SDA ensemble}

\item{inputs}{list of model inputs to use in write.configs.SIPNET}

\item{overwrite}{Default FALSE}

\item{outpath}{if specified, write output to a new directory. Default NULL writes back to the directory being read}
}
\value{
file split up climate file
}
\description{
Splits climate met for SIPNET
}
\author{
Mike Dietze and Ann Raiho
}
