% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_inputs.SIPNET.R
\name{split_sipnet_met}
\alias{split_sipnet_met}
\title{split sipnet clim file based on start and end time}
\usage{
split_sipnet_met(start.time, stop.time, met, overwrite = FALSE, outpath = NULL)
}
\arguments{
\item{start.time}{start date and time for each SDA ensemble}

\item{stop.time}{stop date and time for each SDA ensemble}

\item{met}{path to sipnet clim file to be split}

\item{overwrite}{if \code{TRUE}, overwrite existing target file (Default \code{FALSE})}

\item{outpath}{if specified, write output to a new directory. Default \code{NULL} writes back to the directory being read}
}
\value{
path to split up climate file
}
\description{
split sipnet clim file based on start and end time
}
\author{
Mike Dietze, Ann Raiho, Alexey Shiklomanov
}
