% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg2model.SIPNET.R
\name{veg2model.SIPNET}
\alias{veg2model.SIPNET}
\title{veg2model.SIPNET}
\usage{
veg2model.SIPNET(outfolder, poolinfo, siteid, ens)
}
\arguments{
\item{outfolder}{location to store ncdf files}

\item{poolinfo}{object passed from write_ic contains output from cohort2pool function}

\item{siteid}{object passed from write_ic contains site id}

\item{ens}{number of ensemble members}
}
\value{
result object with filepaths to ncdf files
}
\description{
veg2model.SIPNET
}
\author{
Alexis Helgeson
}
