% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.configs.SIPNET.R
\name{write.config.SIPNET}
\alias{write.config.SIPNET}
\title{Writes a configuration files for SIPNET model}
\usage{
write.config.SIPNET(
  defaults,
  trait.values,
  settings,
  run.id,
  inputs = NULL,
  IC = NULL,
  restart = NULL,
  spinup = NULL
)
}
\arguments{
\item{defaults}{pft}

\item{trait.values}{vector of samples for a given trait}

\item{settings}{PEcAn settings object}

\item{run.id}{run ID}

\item{inputs}{list of model inputs}

\item{IC}{initial condition}

\item{restart}{In case this is a continuation of an old simulation. restart needs to be a list with name tags of runid, inputs, new.params (parameters), new.state (initial condition), ensemble.id (ensemble id), start.time and stop.time.See Details.}

\item{spinup}{currently unused, included for compatibility with other models}
}
\description{
Writes a configuration files for your model
}
\author{
Michael Dietze
}
