\name{write.config.SIPNET}
\alias{write.config.SIPNET}
\title{Writes a configuration files for SIPNET model}
\usage{
  write.config.SIPNET(defaults, trait.values, settings,
    run.id, inputs = NULL, IC = NULL)
}
\description{
  Writes a configuration files for your model
}
\author{
  Michael Dietze
}

