\name{write.restart.SIPNET}
\alias{write.restart.SIPNET}
\title{write.restart.SIPNET}
\usage{
write.restart.SIPNET(out.dir, runid, time, settings, analysis.vec,
  RENAME = TRUE, variables, sample_parameters = FALSE, trait.values = NA,
  met)
}
\arguments{
  \item{out.dir}{output directory}

  \item{runid}{run ID}

  \item{time}{year that is being read}

  \item{settings}{PEcAn settings object}

  \item{analysis.vec}{analysis vector}

  \item{RENAME}{flag to either rename output file or not}

  \item{variables}{}

  \item{sample_parameters}{}

  \item{trait.values}{}

  \item{met}{}
}
\value{
NONE
}
\description{
Write restart files for SIPNET
}
\author{
Ann Raiho \email{araiho@nd.edu}
}

