% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_restart.SIPNET.R
\name{write_restart.SIPNET}
\alias{write_restart.SIPNET}
\title{write_restart.SIPNET}
\usage{
write_restart.SIPNET(outdir, runid, start.time, stop.time, settings,
  new.state, RENAME = TRUE, new.params = FALSE, inputs)
}
\arguments{
\item{outdir}{output directory}

\item{runid}{run ID}

\item{start.time}{start date and time for each SDA ensemble}

\item{stop.time}{stop date and time for each SDA ensemble}

\item{settings}{PEcAn settings object}

\item{new.state}{analysis state vector}

\item{RENAME}{flag to either rename output file or not}

\item{new.params}{list of parameters to convert between different states}

\item{inputs}{list of model inputs to use in write.configs.SIPNET}
}
\value{
NONE
}
\description{
Write restart files for SIPNET
}
\author{
Ann Raiho \email{araiho@nd.edu}
}
